requirements.txt
# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Image Processing
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API requests
requests>=2.31.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
